Update strings.ja-jp.resx
Starting from line 525 (the "Hausdorff-Besikovich" message box text) the strings were in English and not translated. I used the previous commit to search for the translation, but unfortunately I found that there were plenty of Bangla characters in the Japanese text. I used Google Translate to try to figure out the correct translation.

Therefore, this Japanese translation needs to be verified